==> tutorials/micro_cluster_old/30_Analyse_SlurmReal.Rmd <==
---
title: "Analysing SLRUM Real Output and Preparing for Simulation"
output:
  html_document: default
  html_notebook: default
---

```{r setup, echo=TRUE, results="hide",warning=TRUE,message=FALSE}
library(ggplot2)
library(gridExtra)
library(scales)
library(lubridate)
library(stringr)

library(rPython)
library(Rcpp)

#some global locations
top_dir <- "/home/mikola/slurm_simulator3/slurm_sim_tools/validation"

real_top_dir <- "/home/mikola/slurm_simulator3/slurm_real/5"
sim_top_dir <- "/home/mikola/slurm_simulator3/sim/micro3/results"

setwd(top_dir)
source(file.path(top_dir,"../Rutil/trace_job_util.R"))
```

# Reads Data

```{r}
init_trace <- read.csv(file.path(top_dir,"test_trace.csv"))
init_trace$sim_submit <- as.POSIXct(init_trace$sim_submit,format = "%Y-%m-%d %H:%M:%S")
init_trace$sim_dependency <- ""

sacct_r <- read_sacct_out(file.path(real_top_dir,"slurm_acct.out"))
sacct_r$ref_job_id <- as.integer(sub("\\.sh","",sacct_r$JobName))

```

# Which jobs was not done

Sometimes jobs might not run, so lets check did all jobs ran or not.

```{r}
jobs_not_done <- setdiff(init_trace$sim_job_id,sacct_r$ref_job_id)
print(paste("Number of jobs which were not run:",length(jobs_not_done)))
```

```{r}
print(init_trace[init_trace$sim_job_id %in% jobs_not_done,])
```

# Plotting Jobs Submittion and Execution Times

```{r , fig.width=10, fig.height=6}

ggplot(sacct_r)+
    geom_point(aes(x=local_job_id,y=Submit,colour="Submit Time"))+
    geom_segment(aes(x=local_job_id,y=Start,xend=local_job_id,yend=End,colour="Run Time"))+
    scale_colour_manual("",values = c("red","blue", "green"))

```


# Preparing for Simulator
Sometimes some jobs might not run, but we still want to run simulator while waiting for new results 
from real Slurm run. Lets select only jobs which were actually used in real Slurm Run. Here it is all now.

Lets modify sumbit time in init_trace and job_id to set it from real Slurm run and feed it later to simulator.


```{r}

start_time <- min(sacct_r$Submit)
init_trace_start_time <- min(init_trace$sim_submit)

dt<-start_time-init_trace_start_time

init_trace$old_submit <- init_trace$sim_submit
init_trace$sim_submit <- init_trace$old_submit+dt

new_trace <- merge(init_trace,sacct_r,by.x = "sim_job_id", by.y = "ref_job_id", 
                   suffixes = c("",".sacct_r"))
#let the job id be like in real slurm
new_trace$sim_job_id <- new_trace$JobID
submit_not_na <- !is.na(new_trace$Submit)
print(paste("NA submits:",sum(is.na(new_trace$Submit))))
new_trace$sim_submit[submit_not_na]<-new_trace$Submit[submit_not_na]
new_trace$sim_submit_ts <- as.integer(new_trace$sim_submit)
new_trace<-new_trace[order(new_trace$sim_submit_ts),]

new_trace$sim_duration_old <- new_trace$sim_duration
new_trace$sim_duration <- as.integer(unclass(new_trace$End)-unclass(new_trace$Start))

sum(abs(new_trace$sim_duration-new_trace$sim_duration_old))
#lets write traces for simulator
write_trace2(file.path(top_dir,"test_after_slurm_real.trace"),new_trace)



```
