==> tutorials/micro_cluster_old/33_Analyse_Prior.Rmd <==
---
title: "Analysing Priorities from SLRUM Real and Simulated Output"
output: html_notebook
---

```{r setup, echo=TRUE, results="hide",warning=TRUE,message=FALSE}
library(ggplot2)
library(gridExtra)
library(scales)
library(lubridate)
library(stringr)

library(rPython)
library(Rcpp)
library(plyr)

#some global locations
top_dir <- "/home/mikola/slurm_simulator3/slurm_sim_tools/validation"

real_top_dir <- "/home/mikola/slurm_simulator3/slurm_real/s1"
sim_top_dir <- "/home/mikola/slurm_simulator3/sim/micro/results/StartSecondsBeforeFirstJob_30"

setwd(top_dir)
source("../Rutil/trace_job_util.R")

source("micro_conf.R")
```

# Reads Data

```{r}
init_start_time <- as.POSIXct("2017-03-01")


sacct_r0 <- read_sacct_out(file.path(real_top_dir,"slurm_acct.out"),micro_nodes)
dt_r <- min(as.integer(sacct_r0$Submit))-as.integer(init_start_time)
sacct_s0 <- read_sacct_out(file.path(sim_top_dir,"jobcomp.log"),micro_nodes)
dt_s <- min(as.integer(sacct_s0$Submit))-as.integer(init_start_time)
rm(sacct_r0,sacct_s0)
#shift time

print(paste("dt:",dt_r,dt_s))


prio_r <- read.csv(file.path(real_top_dir,"sprio.csv"))
prio_r$t <- as.POSIXct(prio_r$t,format = "%Y-%m-%d %H:%M:%S")
prio_r$t <- prio_r$t -dt_r
prio_s <- read.csv(file.path(sim_top_dir,"sprio.csv"))
prio_s$t <- as.POSIXct(prio_s$t,format = "%Y-%m-%d %H:%M:%S")
prio_s$t <- prio_s$t -dt_s

prio_s$user <- NULL
prio_s$tres <- NULL
prio_s$nice <- NULL
prio_s$partition <- NULL

prio_r$nice <- NULL
prio_r$partition <- NULL

prio_r$Slurm <- "Real"
prio_s$Slurm <- "Simulated"

prio <- rbind(prio_r,prio_s)

```

## Characterize jobs by number of records
```{r}
jobs_rec_r <- count(prio_r,"jobid")
jobs_rec_s <- count(prio_s,"jobid")

jobs_rec <- merge(jobs_rec_r,jobs_rec_s,by="jobid",all=TRUE,suffixes = c("_r","_s"))
jobs_rec <- jobs_rec[order(-jobs_rec$freq_r),]
jobs_rec
```

## Plot selected priorities

```{r , fig.width=5, fig.height=6}
m_job_id<-1269
#prio[prio$jobid==m_job_id,]
p<-grid.arrange(

    ggplot(prio[prio$jobid==m_job_id,],mapping = aes(x=t,y=age,colour=Slurm))+
      geom_line()+ggtitle("A")+xlab("time")+
      theme(legend.position = "none",legend.margin=margin(-15),
          plot.title = element_text(face="bold",hjust = -.17, vjust=0))+
      scale_color_manual(values = c("orangered","royalblue")),
    ggplot(prio[prio$jobid==m_job_id,],mapping = aes(x=t,y=fairshare,colour=Slurm))+
      geom_line()+ggtitle("B")+xlab("time")+
      theme(legend.position = "bottom",legend.justification = "right",legend.margin=margin(-15),
          plot.title = element_text(face="bold",hjust = -.16, vjust=0))+
      scale_color_manual(values = c("orangered","royalblue")),
    ggplot(prio[prio$jobid==m_job_id,],mapping = aes(x=t,y=priority,colour=Slurm))+
      geom_line()+ggtitle("C")+xlab("time")+
      theme(legend.position = "none",legend.margin=margin(-15),
          plot.title = element_text(face="bold",hjust = -.20, vjust=0))+
      scale_color_manual(values = c("orangered","royalblue")),
    nrow=3
)
ggsave('priority.png',p,width = 5,height = 6)
p
```
